first runnable version with generic sword exporter class; further testing needed
